Fixed stylesheets and whatever else, improved install script

+++ install.sh
@@ -9,11 +9,21 @@ qmake ..
 echo "Compiling and linking.."
 make
 
-currentdir=$(pwd)
-echo "Making symbolic link in /usr/bin to $currentdir/tablify.."
-sudo ln -s $currentdir/tablify /usr/bin
-echo "Making config directory.."
-mkdir ~/.config/tablify
-echo "Making empty stylesheet in ~/.config/stylesheet.qss"
-echo "" > ~/.config/tablify/stylesheet.qss
+echo "Making symbolic link in /usr/bin.."
+if [ -f /usr/bin/tablify ]; then
+	echo "/usr/bin/tablify already exists, skipping"
+else
+	currentdir=$(pwd)
+	sudo ln -s $currentdir/tablify /usr/bin
+	echo "Made symbolic link in /usr/bin to $currentdir/tablify.."
+fi
+
+if [ ! -f ~/.config/tablify/stylesheet.qss ]; then
+	echo "~/.config/tablify/stylesheet.qss not found, adding empty file"
+	mkdir ~/.config/tablify
+	echo "" > ~/.config/tablify/stylesheet.qss
+else
+	echo "~/.config/tablify/stylesheet already exists, skipping.."
+fi
+
 echo "Done!"
